libder: add some streaming types
This adds FILE and FD, which are perhaps the more common ones we might see;
the structure was already in-place, just implement the refill functionality
for each one and the callers already anticipate that they might not be able
to collect the entire buffer in one refill call.

This also moves the former fuzz_read into a more general fuzz_stream, that
can either parse out of a buffer or parse from an fd (over a socket) or from
a FILE.  We hack off ~32 bytes from the fuzzing data as a frame to describe
the parameters of a single fuzzing run.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>