Add a README
Explain the non-goals of the project; I really mostly need it for octet/bit
strings, personally, with a couple of OIDs.  I'd strongly prefer to stick
with just dealing with already-encoded payloads that the user can turn into
libder nodes and a build a proper tree to represent the DER output they
would like.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>