libder: allow bit strings and octet strings to be coalesced
Bit strings just need a little bit more logic both because we may have some
unused bits at the end, but also because we need to strip the lead byte from
the subsequent payloads that we're merging in.

We also do a little more validation on bit strings when reading them in to
make sure that we don't have a bogus unused bit or fewer bits than we need
to express the string.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>