libder: actually implement set normalization
I've explored other ways of doing this, but one of my main goals is to not
invalidate existing objects any more than we need to due to normalization.
There are some circumstances where we can't avoid it, e.g., because we need
to coalesce objects into their parents, but ideally one can reason about the
lifetime of a set object.

Using a doubly-linked list for sequences/sets isn't really all that
convenient after all, so just rip that back out.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>