tests: some improvements to our fuzzing setup
Some normalization rules exist now that might leave us with a longer buffer
(e.g., bit string normalization may add an extra byte), so drop the
assertion to the contrary.

Add some constructed bitstrings to our seed corpus generator, and also add
an -H mode that strips the header we use for our streaming fuzzer input so
that we can reuse it for generating a seed corpus for our write fuzzer.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>