libder: overhaul tag storage/reading/writing
It turns out that the identifier has a long form that can also be used, so
we need to account for that as well.  Push the tag out into its own struct
that deconstructs the metadata (class & P/C bits) and either stores it short
or encoded.

We normalize when we read in case some bozo stored a <= 31 tag in the long
form.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>