Add 'derdump', a utility to dump DER encoded files
derdump just takes a series of filenames, and dumps each one to stdout in
order with a header containing the filename.  Files are parsed in verbose
mode, to help locate errors encountered.

While we're here, fix... -ish... output of tags.  The tag class isn't being
encoded in the output, which is kind of important.  Also, go ahead and
handle... -ish... encoded tags in libder_obj_dump(), by just not outputting
the entirely wrong thing.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>