libder: re-read() on EINTR
This isn't the most ideal behavior, but it's good enough for now and fixes
a conflict in the fuzzer with libFuzzer's timeout implementation --
sometimes the resulting SIGALRM would cause an EINTR and we'd assert in the
caller, but ideally we'd be able to handle it a little better.

Ideally, I think we would inform libder_stream_refill() if the read should
be interruptible or not.  If we're interrupted at the very beginning of
attempting to parse an object then we can safely bail out, but the later
refills for pulling the data should likely loop so that we can get a
complete object.

Additionally, it would be nice if we could perhaps bubble up an EINTR to the
caller (likely as LDE_INTR) and let them pull a resumption cookie that can
be used with a new libder_read_resume().  The cookie would effectively
include where we were at in the read state and the stream that we were
reading, which could be complicated.  Notably, we could be a couple of
levels deep into a sequence hierarchy, so we need to be able to resume it
at the previous point, then backtrack up the hierarchy.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>