tests: add a parallel fuzzer
This is a kind of silly one, but add a new fuzzer that is effectively just
fuzz_write with threads.  The main point is to exercise opening and closing
contexts simultaneously in multiple threads to make sure that we don't
break anything as we start tracking open contexts in a global list.  We're
basically abusing libFuzzer to get random access patterns in an attempt to
break it.

I won't be running this one regularly, just if we make a change that seems
suspicious.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>